requirements.txt
about-time==4.2.1
alive-progress==3.1.5
autograd==1.6.2
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
cma==3.2.2
contourpy==1.2.1
cycler==0.12.1
deap==1.4.1
Deprecated==1.2.14
dill==0.3.8
docutils==0.21.2
et-xmlfile==1.1.0
Faker==26.0.0
filelock==3.15.1
fonttools==4.53.0
future==1.0.0
gdown==5.2.0
grapheme==0.6.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.23.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
paretoset==1.2.3
pillow==10.3.0
pyaml==24.4.0
pymoo==0.6.1.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.0
scikit-optimize==0.10.2
scipy==1.13.1
six==1.16.0
soupsieve==2.5
statistics==1.0.3.5
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
